models + bd
