+++ app.js
@@ -1,3 +1,5 @@
+require ('./models/connect');
+
 var createError = require('http-errors');
 var express = require('express');
 var path = require('path');
