+++ models/journey.js
@@ -1,0 +1,13 @@
+var mongoose = require('mongoose');
+
+var journeySchema = mongoose.Schema({
+  departure: String,
+  arrival: String,
+  date: Date,
+  departureTime: String,
+  price: Number,
+});
+
+var journeyModel = mongoose.model('journey', journeySchema);
+
+module.exports = journeyModel;
