+++ models/users
@@ -1,0 +1,12 @@
+var mongoose = require('mongoose');
+
+var userSchema = mongoose.Schema({
+    firstname: String,
+    name: String,
+    email: String,
+    password: String
+})
+
+var userModel = mongoose.model('users', userSchema);
+
+module.exports = userModel;
