+++ package.json
@@ -11,7 +11,7 @@
     "ejs": "~2.5.7",
     "express": "~4.16.0",
     "http-errors": "~1.6.2",
-    "mongoose": "^5.7.8",
+    "mongoose": "^5.13.14",
     "morgan": "~1.9.0",
     "nodemon": "^2.0.18"
   }
